pyproject.toml
[tool.poetry]
name = "missaocertificacaomundo1"
version = "0.1.0"
description = ""
authors = ["Wellington Fonseca <wellington.cesar.fonseca@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
pendulum = "^2.1.2"
mypy = "^1.4.1"
python-dotenv = "^1.0.0"
rich = "^13.5.2"
validate-docbr = "^1.10.0"
unidecode = "^1.3.6"
django-debug-toolbar = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
